requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
pinecone-client
tiktoken
unstructured[pdf]
python-magic
tavily-python
tenacity
